# utils: speed up timer by embedding the time.Timer