Note a huge speedup, but we're resetting timers _a lot_ since we're using it to pace packets.
```
name      old time/op  new time/op  delta
Timer-16  77.9ns ± 4%  75.8ns ± 3%  -2.59%  (p=0.000 n=50+47)
```