@onee-only 2024-05-19
Seems like `(*time.Timer).Reset` does not work for timer as value for some reason.

Since the timer is initialized with `math.MaxInt64` duration, and is not reset due to the problem above,
the tests are being timed out waiting for receiving from the timer's channel.

I've tested behavior with code snippet below.

using timer as pointer:
```go
func main() {
	now := time.Now()
	t := time.NewTimer(time.Second)
	t.Reset(time.Second * 2)

	<-t.C
	log.Println(time.Since(now))
	// 2024/05/19 17:09:14 2.000662625s
}
```

using timer as value:
```go
func main() {
	now := time.Now()

	t := *time.NewTimer(time.Second)
	t.Reset(time.Second * 2)

	<-t.C
	log.Println(time.Since(now))
	// 2024/05/19 17:07:54 1.001052084s
}
```

I'm not sure if there is a workaround.