@sukunrt 2024-05-19
Can you try embedding a pointer to the timer?