@marten-seemann 2024-05-19
> Can you try embedding a pointer to the timer?

That's what we currently have, or am I misunderstanding what you're suggesting?