@sukunrt 2024-05-19
You aren't. Embedding the timer is what we have currently. I think this is what's happening. 

You cannot copy a time.Timer{}. Currently, we create a new timer using `time.NewTimer` which returns a pointer to the created timer. Now we dereference this pointer and copy the underlying timer value into our struct. This also copies the `time.Timer` struct's underlying `runtimeTimer`. This copying is illegal. The way runtime tracks these timers is by their address which it stores when creating the timer with `time.NewTimer`, the address of our struct is different because we copied this.  

Now when we call `timer.Reset` on the copied timer the implementation goes through the reset steps, but this new value will never be used because the runtime isn't tracking this copied timer at all. Specifically this slice doesn't have the copied timer: 
https://github.com/golang/go/blob/go1.22.3/src/runtime/proc.go#L3974